% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deft_show.R
\name{deft_show}
\alias{deft_show}
\title{Show deft result}
\usage{
deft_show(deft, element, study_labels = NULL,
  measure_label = "Hazard ratio", trans = base::exp,
  show_model = ifelse(element == "all", FALSE, TRUE),
  show_stats = list(`I^2` = rlang::quo(sprintf("\%0.1f\%\%", I2)), p =
  rlang::quo(format.pval(QEp, digits = 2))), ...)
}
\arguments{
\item{deft}{result from \link{deft_do}.}

\item{element}{'all' or 'subgroup'.}

\item{study_labels}{labels for studies.}

\item{measure_label}{label for meature, e.g. 'HR'.}

\item{trans}{an optional transform function used on the numeric data for plotting the axes}

\item{show_model}{if `TRUE`, show model result, otherwise only show forest plot for studies.}

\item{show_stats}{a \code{list} of stats to show at the bottom of the forest plot for e.g. heterogeneity}

\item{...}{other arguments except 'panels', 'trans', 'study_labels',
and 'show_stats' passed to \code{\link[forestmodel:forest_rma]{forestmodel::forest_rma()}}.}
}
\value{
a \code{ggplot} object
}
\description{
Show deft result
}
\examples{
data('wang2019')
}
