% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{easy_forest}
\alias{easy_forest}
\alias{easy_forest.rma}
\title{easy forest plots}
\usage{
easy_forest()

\method{easy_forest}{rma}(x, annotate = TRUE, addfit = TRUE,
  addcred = FALSE, showweights = FALSE, xlim, alim, clim, ylim, at,
  steps = 5, level = x$level, refline = 0, digits = 2L, width, xlab,
  slab, mlab, ilab, ilab.xpos, ilab.pos, order, transf, atransf, targs, rows,
  efac = 1, pch = 15, psize, col, border, lty, cex, cex.lab, cex.axis,
  annosym, ...)
}
\arguments{
\item{x}{an object of class \code{"rma"}.}

\item{annotate}{logical specifying whether annotations should be added to the plot (the default is \code{TRUE}).}

\item{addfit}{logical specifying whether the summary estimate (for models without moderators) or fitted values (for models with moderators) should be added to the plot. See \sQuote{Details}.}

\item{addcred}{logical specifying whether the bounds of the credibility/prediction interval should be added to the plot. See \sQuote{Details}.}

\item{showweights}{logical specifying whether the annotations should also include the weights given to the observed effects or outcomes during the model fitting (the default is \code{FALSE}).}

\item{xlim}{horizontal limits of the plot region. If unspecified, the function tries to set the horizontal plot limits to some sensible values.}

\item{alim}{the actual x-axis limits. If unspecified, the function tries to set the x-axis limits to some sensible values.}

\item{clim}{limits for the confidence/credibility/prediction intervals. If unspecified, no limits are used.}

\item{ylim}{the y limits of the plot. If unspecified, the function tries to set the y-axis limits to some sensible values.}

\item{at}{position of the x-axis tick marks and corresponding labels. If unspecified, the function tries to set the tick mark positions/labels to some sensible values.}

\item{steps}{the number of tick marks for the x-axis (the default is 5). Ignored when the user specifies the positions via the \code{at} argument.}

\item{level}{numerical value between 0 and 100 specifying the confidence interval level (the default is to take the value from the object).}

\item{refline}{value at which a vertical \sQuote{reference} line should be drawn (the default is 0). The line can be suppressed by setting this argument to \code{NA}.}

\item{digits}{integer specifying the number of decimal places to which the tick mark labels of the x-axis and the annotations should be rounded (the default is \code{2L}). Can also be a vector of two integers, the first specifying the number of decimal places for the annotations, the second for the x-axis labels. When specifying an integer (e.g., \code{2L}), trailing zeros after the decimal mark are dropped for the x-axis labels. When specifying a numerical value (e.g., \code{2}), trailing zeros are retained.}

\item{width}{optional integer to manually adjust the width of the columns for the annotations.}

\item{xlab}{title for the x-axis. If unspecified, the function tries to set an appropriate axis title.}

\item{slab}{optional vector with labels for the \eqn{k} studies. If unspecified, the labels are either taken from the object (if study labels were specified) or simple labels are created within the function. To suppress labels, set this argument to \code{NA}.}

\item{mlab}{optional character string giving a label to the summary estimate from a fixed- or random-effects model. If unspecified, the label is created within the function.}

\item{ilab}{optional vector, matrix, or data frame providing additional information about the studies that should be added to the plot.}

\item{ilab.xpos}{vector of numerical value(s) specifying the x-axis position(s) of the variable(s) given via \code{ilab} (must be specified if \code{ilab} is specified).}

\item{ilab.pos}{integer(s) (either 1, 2, 3, or 4) specifying the alignment of the vector(s) given via \code{ilab} (2 means right, 4 mean left aligned). If unspecified, the default is to center the labels.}

\item{order}{optional character string specifying how the studies should be ordered. See \sQuote{Details}.}

\item{transf}{optional argument specifying the name of a function that should be used to transform the observed effect sizes, summary estimates, fitted values, and confidence interval bounds (e.g., \code{transf=exp}; see also \link{transf}). If unspecified, no transformation is used.}

\item{atransf}{optional argument specifying the name of a function that should be used to transform the x-axis labels and annotations (e.g., \code{atransf=exp}; see also \link{transf}). If unspecified, no transformation is used.}

\item{targs}{optional arguments needed by the function specified via \code{transf} or \code{atransf}.}

\item{rows}{optional vector specifying the rows (or more generally, the horizontal positions) for plotting the outcomes. If unspecified, the function sets this value automatically. Can also be a single value specifying the row (horizontal position) of the first outcome (the remaining outcomes are then plotted below this starting row).}

\item{efac}{vertical expansion factor for confidence interval limits, arrows, and the symbol used to denote summary estimates. The default value of 1 should usually work okay. Can also be a vector of two numbers, the first for CI limits and arrows, the second for summary estimates. Can also be a vector of three numbers, the first for CI limits, the second for arrows, the third for summary estimates.}

\item{pch}{plotting symbol to use for the observed effect sizes or outcomes. By default, a filled square is used. See \code{\link{points}} for other options. Can also be a vector of values.}

\item{psize}{optional vector with point sizes for the observed effect sizes or outcomes. If unspecified, the point sizes are drawn proportional to the precision of the estimates.}

\item{col}{optional character string specifying the name of a color to use for the summary polygon or fitted values. If unspecified, the function sets a default color.}

\item{border}{optional character string specifying the name of a color to use for the border of the summary polygon or fitted values. If unspecified, the function sets a default color.}

\item{lty}{optional character string specifying the line type for the confidence intervals (if unspecified, the function sets this to \code{"solid"} by default).}

\item{cex}{optional character and symbol expansion factor. If unspecified, the function tries to set this to a sensible value.}

\item{cex.lab}{optional expansion factor for the x-axis title. If unspecified, the function tries to set this to a sensible value.}

\item{cex.axis}{optional expansion factor for the x-axis labels. If unspecified, the function tries to set this to a sensible value.}

\item{annosym}{optional vector of length 3 to change the left bracket, separation, and right bracket symbols for the annotations.}

\item{...}{other arguments.}
}
\description{
A wrapper of forest function in metafor package.
More easier use for subgroup analysis.
}
\details{
This is a generic function build on forest function in metafor
package. Even though the forest function is flexible, it is not very easy
generate plots readily for publication, especially in subgroup analysis.
}
\author{
Shixiang Wang <w_shixiang@163.com>

Shixiang Wang <w_shixiang@163.com>
}
