% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{log_ES}
\alias{log_ES}
\title{log transformation for effect size estimation according to confidence level and distribution}
\usage{
log_ES(data, distribution = c("N", "t", "F"), conf_level = 0.05,
  df1 = Inf, df2 = NULL)
}
\arguments{
\item{data}{a dataframe include effect size, ci.lb, ci.ub in the first, second and third columns}

\item{distribution}{a character specify}
}
\description{
A variety of different outcome measures which used in meta-analysis,
as input, are in the form of log, such as Hazard Ratio. This function
is used to do log transformation to calculate effect size and
standard error. Then the result can be easier used for model fit.
}
\examples{
hr <- c(3.12, 1.15) # specify hazard ratios (hr)
ci.lb <- c(2.2, 1.03) # specify lower bound for hr confidence intervals
ci.ub <- c(4.1, 2.6) # specify upper bound for hr confidence intervals
dat = data.frame(hr=hr, ci.lb=ci.lb, ci.ub=ci.ub)
log_ES(dat)

}
\author{
Shixiang Wang <w_shixiang@163.com>
}
