% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wang2019}
\alias{wang2019}
\title{Hazard ratio (HR) for disease progression analysis comparing
TMB-high with TMB-low in three NSCLC datasets}
\format{a \code{data.frame}}
\source{
Wang, Shixiang, et al. "The predictive power of tumor mutational burden
in lung cancer immunotherapy response is influenced by patients' sex."
International journal of cancer (2019).
}
\description{
Hazard ratio (HR) for disease progression analysis comparing
TMB-high with TMB-low in three NSCLC datasets
}
\examples{
data("wang2019")
}
